feat: 해시/브라우저 라우터 공통 구조로 통합
