+++ src/features/auth/index.js
@@ -1,0 +1,6 @@
+import LoginPage from "./LoginPage";
+import { AuthStore } from "./AuthStore";
+import { bindLoginEvent } from "./LoginEvents";
+
+export default LoginPage;
+export { AuthStore, bindLoginEvent };
