+++ src/features/main/MainPage.js
@@ -1,5 +1,5 @@
-import Header from "../components/Footer";
-import Footer from "../components/Header";
+import Header from "../../shared/components/Header";
+import Footer from "../../shared/components/Footer";
 
 const MainPage = () => `
   <div class="bg-gray-100 min-h-screen flex justify-center">
