+++ src/features/main/index.js
@@ -1,0 +1,3 @@
+import MainPage from "./MainPage";
+
+export default MainPage;
