+++ src/features/profile/index.js
@@ -1,0 +1,5 @@
+import ProfilePage from "./ProfilePage";
+import { bindProfileEvent } from "./profileEvents";
+
+export default ProfilePage;
+export { bindProfileEvent };
