+++ src/main.hash.js
@@ -1,1 +1,5 @@
-import "./main.js";
+import { initializeApp } from "./app";
+
+const IS_HASH_MODE = true;
+
+initializeApp(IS_HASH_MODE);
